Improved view-model abstraction
