+++ app/src/main/java/org/listenbrainz/android/viewmodel/FollowUnfollowModel.kt
@@ -1,0 +1,73 @@
+package org.listenbrainz.android.viewmodel
+
+import kotlinx.coroutines.CoroutineDispatcher
+import kotlinx.coroutines.withContext
+import org.listenbrainz.android.di.IoDispatcher
+import org.listenbrainz.android.model.ResponseError
+import org.listenbrainz.android.model.User
+import org.listenbrainz.android.repository.social.SocialRepository
+import org.listenbrainz.android.util.Resource
+import kotlin.coroutines.CoroutineContext
+
+abstract class FollowUnfollowModel<UiState>(
+    private val repository: SocialRepository,
+    @IoDispatcher private val ioDispatcher: CoroutineDispatcher
+): BaseViewModel<UiState>() {
+    
+    /**
+     * @param [invertUiState] **should** check for cancellation internally.*/
+    suspend fun CoroutineContext.optimisticallyFollowUser(
+        user: User,
+        index: Int,
+        invertUiState: CoroutineContext.(index: Int) -> Unit
+    ) {
+        /** Determines if user is followed or not.*/
+        fun userIsAlreadyFollowed(error: ResponseError?): Boolean {
+            return error == ResponseError.BAD_REQUEST &&
+                    error.actualResponse?.contains("already following") == true
+        }
+        
+        invertUiState(index)
+        
+        val result = withContext(ioDispatcher) {
+            repository.followUser(user.username)
+        }
+        when (result.status) {
+            Resource.Status.FAILED -> {
+                emitError(result.error)
+                
+                if (userIsAlreadyFollowed(result.error)){
+                    // We won't toggle back follow state if user is already followed.
+                    return
+                }
+                
+                invertUiState(index)
+            }
+            else -> Unit
+        }
+    }
+    
+    /**
+     * @param [invertUiState] **should** check for cancellation internally.*/
+    suspend fun CoroutineContext.optimisticallyUnfollowUser(
+        user: User,
+        index: Int,
+        invertUiState: CoroutineContext.(index: Int) -> Unit
+    ) {
+        
+        invertUiState(index)
+        
+        val result = withContext(ioDispatcher) {
+            repository.unfollowUser(user.username)
+        }
+        return when (result.status) {
+            Resource.Status.FAILED -> {
+                // Since same response is given by server even if user is unfollowed or not, we
+                // won't do anything here.
+                invertUiState(index)
+                emitError(result.error)
+            }
+            else -> Unit
+        }
+    }
+}
